+++ spec/rubocop/cop/lint/number_conversion_spec.rb
@@ -77,5 +77,35 @@ RSpec.describe RuboCop::Cop::Lint::NumberConversion do
         Complex('10')
       RUBY
     end
+
+    it 'when using Time' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        Time.now.to_i
+        Time.now.to_f
+        Time.strptime("2000-10-31", "%Y-%m-%d").to_i
+        Time.httpdate("Thu, 06 Oct 2011 02:26:12 GMT").to_f
+        Time.now.to_datetime.to_i
+      RUBY
+    end
+
+    it 'when using DateTime' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        DateTime.new(2012, 8, 29, 22, 35, 0).to_i
+        DateTime.new(2012, 8, 29, 22, 35, 0).to_f
+      RUBY
+    end
+
+    it 'when using Time/DateTime with multiple method calls' do
+      expect_no_offenses(<<-RUBY.strip_indent)
+        Time.now.to_datetime.to_i
+        DateTime.civil(2005, 2, 21, 10, 11, 12, Rational(-6, 24)).utc.to_f
+        Time.zone.now.to_datetime.to_f
+        DateTime.new(2012, 8, 29, 22, 35, 0)
+                .change(day: 1)
+                .change(month: 1)
+                .change(year: 2020)
+                .to_i
+      RUBY
+    end
   end
 end